Fix build

+++ src/sock.c
@@ -503,8 +503,8 @@ static void mg_iotest(struct mg_mgr *mgr, int ms) {
 #elif MG_ENABLE_POLL
   nfds_t n = 0;
   for (struct mg_connection *c = mgr->conns; c != NULL; c = c->next) n++;
-  int pollCount = n == 0 ? 1 : 1; // Avoid zero-length VLA
-  int pollSize = pollCount * (int) sizeof(struct pollfd);
+  size_t pollCount = n == 0 ? 1 : 1; // Avoid zero-length VLA
+  size_t pollSize = pollCount * sizeof(struct pollfd);
   struct pollfd *fds = alloca(pollSize);
 
   memset(fds, 0, pollSize);
